BUG: 212855 BUG: 156296 Added two files for the fix. protocols/yahoo/libkyahoo/alivetask.h protocols/yahoo/libkyahoo/alivetask.cpp
The code from ping.h and cpp now actually resides in those files,
and the ping task actually sends the ping not the alive message,
Alive messages are sent to keep communcation between Yahoo and the client,
but the Ping keeps you connected to the service so as they dont send a disconnect after the hour.

I have set the time out to 59 minutes for the alive and the alive to 1 minute.

svn path=/branches/KDE/4.4/kdenetwork/kopete/; revision=1079333