Do not reset m_clients when the video-device is changed.
This fixes a problem that comes up when the video-device is switched in the config-dialog while a protocol is capturing from it:
Capturing (by the protocol) stops working when the config-dialog is quit.

Also check if the previous device has been closed successfully before opening the new device.

Backport of r1081461

svn path=/branches/KDE/4.4/kdenetwork/kopete/; revision=1081464