Implement accepting/refusing/requesting presence authorization from the contact context menu
This enables us to recover from eg. (perhaps accidentally) having declined presence authorization or just closing the contact popup, which would otherwise leave us with an useless contact. Only the actions which make sense in the current state are enabled, which serves as additional display on the subscription/publishing state.

svn path=/trunk/playground/network/kopete/protocols/telepathy/; revision=1100853