Automatically adjust Jabber priority based on the presence state.
From the original patch description:
"This patch adds the ability to adjust the priority of a jabber resource when
going to an "away" state. This can be used to "redirect" messages to a
different resource (e.g an IM client on a mobile device) base on the presence
state of kopete."

See http://reviewboard.kde.org/r/4274/

Patch from Ralf Haferkamp, thanks!

FEATURE: 97998
GUI:
CCMAIL: rhafer@suse.de

svn path=/trunk/KDE/kdenetwork/kopete/; revision=1145738