Per metacontact statistics DB is created only when contact goes online or when statistics for an offline metacontact is requested. This improves plugin startup and shutdown times and also helps with UI responsiveness.
Now shutdown time is proportional to the number of metacontacts that went
online since Kopete has started plus the number of offline metacontacts
that the user requested statistics from. That is better then being
proportional to the number of all metacontacts registered but is not that
good yet.

BUG: 117989, 138903
CCBUG: 246785
FIXED-IN: 4.6.3

svn path=/trunk/KDE/kdenetwork/kopete/; revision=1227533