Fix the build with giflib 4.2.0+
giflib 4.2.0 removed the PrintGifError() function from the main
library and has started providing GifError() and GifErrorString()
instead.

We can't just switch to that because of libungif and older giflib
releases. Check at configure-time whether GifErrorString() is present
in libgif and, if it is, use that; otherwise, continue using
PrintGifError().

I do not like adding more ADD_DEFINITION() calls, but in this case I
preferred to make as few changes as possible.

CCMAIL: kde-packager@kde.org


svn path=/trunk/KDE/kdenetwork/kopete/; revision=1298170