Automatically logout and login with suspend/resume
This patch logs off each account when PC is suspended and restores the online state of each account on resume.
This patch also addresses another problem of kopete: When the user was setting kopete to "Offline" and the network connection was lost and recovered kopete would reset all accounts to "Online".

Thanks Benni Hill for patch!

BUG: 212037
BUG: 149544
FEATURE: 149544
FIXED-IN: 4.11
REVIEW: 6511

svn path=/trunk/KDE/kdenetwork/kopete/; revision=1355161