Fix function XmlContactStorage::checkGroupIds() in contactlist
This function does not expect that Temporary or Top Level group can be already
in group list. When this happened, function changing all internal group ids
and Top Level group has id 1. All ids are unique. So it also break test
xmlcontactstorage_test.

This patch fixing this problem and do not chaning group ids when all are unique.
So this also fix test xmlcontactstorage_test.

CC: aacid@kde.org