Fix formatting messages (both plain and html) when OTR plugin is loaded but not used
Before this patch OTR plugin escaped all plain text messages even if it did not
encrypted them. So it was not able to send message with char '<'. Even worse OTR
plugin converted all HTML messages (also those which was not encrypted) into
plain text with all HTML tags. So both problems caused that it was not possible
to have OTR plugin enabled and not using it (e.g. for contacts which did not
supported OTR) because it manged all messages.

Because OTR messages support HTML, every message which is OTR encoded should be
converted to HTML first. But messages which will not be OTR encoded should not
be converted to HTML as other side does not have to support HTML messages.

This patch fixing problem with mangling plain text and HTML messages which are
not going to be encrypted and also fixing escaping (converting to html) plain
text messages.