Fix crash on chat window closing
Fix crash on chat window closing which is obviously caused by
reparenting KHTMLPart (which is also KPart) which in turn leads
to storing pointers to both parents, old and new, in KPartManager
and crashing the application when manager tries to delete them
in parallel with QWidget tree of the window itself.

It's OK to leave KHTMLPart's view without parent on creation as
it instantly gets parented when inserting into QSplitter.