When updating OTR GUI icon properly set OTR instance tag
Without configured instance tag libotr library does not encrypt sent
messages and moreover it even does not report any error that message was
not encrypted.

This should fix a bug when OTR "encrypted" icon is shown in GUI and libotr
itself does not want to encrypt messages. It happened when Kopete window
with active OTR session was closed and after that again opened.

BUG: 362535
FIXED-IN: 16.12