Fix jabber support for XEP-0022 and XEP-0184 for Message Delivery
Message Delivery ability depends on jabber client. To one jabber account it
is possible to connect from multiple clients at same time. So caching
information if particular contact support Message Delivery must be done for
jabber resource, not for whole jabber contact.

This patch fixes bug in which Kopete shown animated icon for message
delivery forever. It happened in case Kopete sent message to jabber
resource which do not support XEP-0022 or XEP-0184 but for that contact
there is resource with XEP-0022 or XEP-0184 support.